+++ dginn/core.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 import numpy as np
 import tensorflow as tf
 
-from utils import *
+from dginn.utils import *
 
 # this should be only in the call module, all other modules should not have it!!!
 # best keep it in the main fx!
